Moved SW360ClientException to a sub package of client.
Issue: eclipse#423

All classes that are going to be part of the SW360 client library will
be located in the client namespace.
