+++ modules/sw360/sw360-client/src/main/java/org/eclipse/sw360/antenna/sw360/rest/resource/attachments/SW360Attachment.java
@@ -13,7 +13,7 @@ package org.eclipse.sw360.antenna.sw360.rest.resource.attachments;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import org.eclipse.sw360.antenna.sw360.rest.SW360SimpleHalResource;
-import org.eclipse.sw360.antenna.sw360.utils.SW360ClientException;
+import org.eclipse.sw360.antenna.sw360.client.utils.SW360ClientException;
 
 import java.nio.file.Path;
 import java.util.Optional;
